requirements.txt
# Intent Classifier Service - Dependencies
# Flask + Gunicorn backend with OCI GenAI Embedding + GradientBoosting

# Web Framework
Flask>=3.1.2,<3.2
python-dotenv>=1.2.1,<2
gunicorn>=23.0.0,<25.0.0
gevent>=25.9.1,<26

# OCI SDK (GenAI Embedding)
oci>=2.167.1,<3

# Machine Learning
scikit-learn>=1.7.2,<1.8.0
numpy>=2.0.2,<2.1.0

# HTTP Client (for OCI API calls)
requests>=2.32.5,<3

# Job scheduling / caching utilities
cachetools>=6.2.2,<7

# Excel file processing
openpyxl>=3.1.5,<4

# Logging
python-json-logger>=4.0.0,<5

# Security
python-jose[cryptography]>=3.5.0,<4
# Optional observability dependency. Install only when exposing /api/v1/metrics.
prometheus-client>=0.24.1,<0.25

# Development & Testing (optional)
pytest>=9.0.2,<10; python_version >= "3.10"
pytest>=8.3.3,<9; python_version < "3.10"
pytest-cov>=7.0.0,<8
ruff>=0.15.1,<0.16
black>=26.1.0,<27; python_version >= "3.10"
black>=25.1.0,<26; python_version < "3.10"
mypy>=1.19.1,<2
bandit>=1.9.3,<2; python_version >= "3.10"
bandit>=1.8.6,<1.10; python_version < "3.10"
pip-audit>=2.10.0,<3; python_version >= "3.10"
pip-audit>=2.9.0,<2.10; python_version < "3.10"
